refactor(product-selection): change coordinate types from Double to BigDecimal
- Update ChosenLocation aggregate to use BigDecimal for latitude/longitude
- Update all related DTOs, commands, and events
- Refactor GeoDistance utility to use shared GeoDistanceCalculator

+++ product-selection-service/src/main/kotlin/org/pv293/kotlinseminar/productSelectionService/events/impl/LocationChosenEvent.kt
@@ -1,9 +1,10 @@
 package org.pv293.kotlinseminar.productSelectionService.events.impl
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class LocationChosenEvent(
     val locationId: UUID,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: BigDecimal,
+    val longitude: BigDecimal,
 )
